drm/i915: Disable verbose state checks
For stable releases, we disable the verbose state checks in the i915 driver.
They are mostly for development purposes and they fill the retrace server
to the point of uselessness.  We leave these enabled in rawhide kernels,
where they are both more likely to get attention from upstream and more
likely to be fixable.

Signed-off-by: Josh Boyer <jwboyer@fedoraproject.org>